+++ src/main/kotlin/ua/pp/lumivoid/ClientOptions.kt
@@ -1,5 +1,6 @@
 package ua.pp.lumivoid
 
+import net.minecraft.item.ItemStack
 import net.minecraft.util.math.BlockPos
 import ua.pp.lumivoid.util.features.AutoWire
 
@@ -8,5 +9,7 @@ object ClientOptions {
     var autoWireMode: AutoWire = AutoWire.AUTO_REDSTONE
     var autoWireLastBlock: BlockPos = BlockPos(0, 0, 0)
     var isAirPlaceEnabled: Boolean = false
+    var rememberLastSelectedItems: ItemStack? = null
+    var illegalFeatureNotified: Boolean = false
     //var autoWireBlockCounter: Int = 1
 }