+++ src/main/kotlin/ua/pp/lumivoid/commands/RedstoneHelperCommand.kt
@@ -4,7 +4,6 @@ package ua.pp.lumivoid.commands
 
 import com.mojang.brigadier.CommandDispatcher
 import com.mojang.brigadier.arguments.BoolArgumentType
-import com.mojang.brigadier.arguments.IntegerArgumentType
 import com.mojang.brigadier.arguments.StringArgumentType
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource
@@ -26,21 +25,9 @@ object RedstoneHelperCommand {
         dispatcher.register(ClientCommandManager.literal("redstone-helper")
             .executes { context ->
                 logger.debug("/redstone-helper: Missing arguments!")
-                context.source.sendError(Text.translatable("redstone-helper.stuff.info.error.missing_arguments"))
+                context.source.sendError(Text.translatable(Constants.LOCALIZEIDS.STUFF_INFO_ERROR_MISSINGARGUMENTS))
                 1
             }
-            .then(ClientCommandManager.literal("help")
-                .executes { context ->
-                    context.source.sendFeedback(Text.translatable("redstone-helper.help.help"))
-                    1
-                }
-            )
-            .then(ClientCommandManager.literal("version")
-                .executes { context ->
-                    context.source.sendFeedback(Text.translatable("redstone-helper.help.version", Constants.MOD_VERSION))
-                    1
-                }
-            )
             .then(ClientCommandManager.literal("check-updates")
                 .executes { context ->
                     val checkerText = VersionChecker.checkRedstoneHelperVersionLocalized(
@@ -49,8 +36,8 @@ object RedstoneHelperCommand {
                         checkForAlpha = true
                     )
                     context.source.sendFeedback(checkerText)
-                    if (checkerText != Text.translatable("redstone-helper.stuff.version_checker.up_to_date") && checkerText != Text.translatable("redstone-helper.stuff.version_checker.unable_to_check_version")) {
-                        context.source.sendFeedback(Text.translatable("redstone-helper.stuff.version_checker.skip_version"))
+                    if (checkerText != Text.translatable(Constants.LOCALIZEIDS.STUFF_VERSIONCHECKER_UPTODATE) && checkerText != Text.translatable(Constants.LOCALIZEIDS.STUFF_VERSIONCHECKER_UNABLETOCHECKVERSION)) {
+                        context.source.sendFeedback(Text.translatable(Constants.LOCALIZEIDS.STUFF_VERSIONCHECKER_SKIPVERSION))
                     }
                     1
                 }
@@ -69,7 +56,7 @@ object RedstoneHelperCommand {
                 )
                 .then(ClientCommandManager.literal("skip")
                     .executes { context ->
-                        context.source.sendFeedback(Text.translatable("redstone-helper.stuff.version_checker.skipping"))
+                        context.source.sendFeedback(Text.translatable(Constants.LOCALIZEIDS.STUFF_VERSIONCHECKER_SKIPPING))
                         VersionChecker.skipVersion()
                         VersionChecker.checkRedstoneHelperVersionWithToast()
                         1
@@ -77,7 +64,7 @@ object RedstoneHelperCommand {
                 )
                 .then(ClientCommandManager.literal("clear-skips")
                     .executes { context ->
-                        context.source.sendFeedback(Text.translatable("redstone-helper.stuff.version_checker.clear_skips"))
+                        context.source.sendFeedback(Text.translatable(Constants.LOCALIZEIDS.STUFF_VERSIONCHECKER_CLEARSKIPS))
                         VersionChecker.clearSkippedVersion()
                         VersionChecker.checkRedstoneHelperVersionWithToast()
                         1
@@ -87,7 +74,7 @@ object RedstoneHelperCommand {
             .then(ClientCommandManager.literal("notification")
                 .executes { context ->
                     logger.debug("/redstone-helper: Missing arguments!")
-                    context.source.sendError(Text.translatable("redstone-helper.stuff.info.error.missing_arguments"))
+                    context.source.sendError(Text.translatable(Constants.LOCALIZEIDS.STUFF_INFO_ERROR_MISSINGARGUMENTS))
                     1
                 }
                 .then(ClientCommandManager.literal("clear-queue")
@@ -99,13 +86,13 @@ object RedstoneHelperCommand {
                 .then(ClientCommandManager.literal("add-to-queue")
                     .executes { context ->
                         logger.debug("/redstone-helper: Missing arguments!")
-                        context.source.sendError(Text.translatable("redstone-helper.stuff.info.error.missing_arguments"))
+                        context.source.sendError(Text.translatable(Constants.LOCALIZEIDS.STUFF_INFO_ERROR_MISSINGARGUMENTS))
                         1
                     }
                     .then(ClientCommandManager.argument("short", BoolArgumentType.bool())
                         .executes { context ->
                             logger.debug("/redstone-helper: Missing arguments!")
-                            context.source.sendError(Text.translatable("redstone-helper.stuff.info.error.missing_arguments"))
+                            context.source.sendError(Text.translatable(Constants.LOCALIZEIDS.STUFF_INFO_ERROR_MISSINGARGUMENTS))
                             1
                         }
                         .then(ClientCommandManager.argument("text", StringArgumentType.greedyString())
@@ -118,28 +105,6 @@ object RedstoneHelperCommand {
                 )
             )
             .then(ClientCommandManager.literal("test")
-                .then(ClientCommandManager.literal("notification")
-                    .then(ClientCommandManager.argument("count", IntegerArgumentType.integer())
-                        .then(ClientCommandManager.argument("short", BoolArgumentType.bool())
-                            .executes { context ->
-                                context.source.sendFeedback(Text.literal("Testing notification"))
-                                for (i in 1..IntegerArgumentType.getInteger(context, "count")) {
-                                    HudToast.addToastToQueue(Text.literal(" Test notification $i\n Hello World!\nYour ad here"), BoolArgumentType.getBool(context, "short"))
-                                }
-                                1
-                            }
-                        )
-                    )
-                )
-                .then(ClientCommandManager.literal("raw-notification")
-                    .then(ClientCommandManager.argument("short", BoolArgumentType.bool())
-                        .executes { context ->
-                            context.source.sendFeedback(Text.literal("Testing raw notification"))
-                            HudToast.showToast(Text.literal("Test raw notification"), BoolArgumentType.getBool(context, "short"))
-                            1
-                        }
-                    )
-                )
                 .then(ClientCommandManager.literal("help")
                     .executes {
                         logger.debug("/redstone-helper: Opening manual menu")