+++ src/main/kotlin/ua/pp/lumivoid/gui/AutowireScreen.kt
@@ -6,6 +6,7 @@ import io.wispforest.owo.ui.component.CheckboxComponent
 import io.wispforest.owo.ui.component.DropdownComponent
 import io.wispforest.owo.ui.component.LabelComponent
 import io.wispforest.owo.ui.container.FlowLayout
+import io.wispforest.owo.ui.core.Color
 import io.wispforest.owo.ui.core.Component
 import io.wispforest.owo.ui.core.Surface
 import net.minecraft.text.Text
@@ -13,6 +14,7 @@ import net.minecraft.util.Identifier
 import ua.pp.lumivoid.ClientOptions
 import ua.pp.lumivoid.Config
 import ua.pp.lumivoid.Constants
+import ua.pp.lumivoid.util.Calculate
 import ua.pp.lumivoid.util.features.AutoWire
 
 class AutowireScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, DataSource.asset(Identifier.of(Constants.MOD_ID, "autowire_ui_model"))) {
@@ -30,6 +32,18 @@ class AutowireScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, Data
             layout.surface(Surface.DARK_PANEL)
         } else {
             layout.surface(Surface.PANEL)
+
+            val color: Color
+            Calculate.hexToRGB(0x3F3F3F).let { (r, g, b, a) ->
+                color = Color(r, g, b, a)
+            }
+            layout.children().forEach { component ->
+                logger.info("child")
+                if (component is LabelComponent && !component.text().string.contains("__colored")) {
+                    logger.info(component.text().string)
+                    component.color(color)
+                }
+            }
         }
         if (Config().enableBackgroundBlur) {
             rootComponent.surface(Surface.blur(100F, 10F))
@@ -38,7 +52,7 @@ class AutowireScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, Data
         }
 
         autowireState.checked(ClientOptions.isAutoWireEnabled)
-        currentMode.text(Text.translatable("redstone-helper.feature.auto_wire.current_auto_wire_mode", Text.translatable("redstone-helper.feature.auto_wire.modes.${ClientOptions.autoWireMode.toString().lowercase()}")))
+        currentMode.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_AUTOWIRE_CURRENTAUTOWIREMODE, Text.translatable("redstone-helper.feature.auto_wire.modes.${ClientOptions.autoWireMode.toString().lowercase()}")))
 
         autowireState.onChanged {
             ClientOptions.isAutoWireEnabled = autowireState.isChecked
@@ -59,7 +73,7 @@ class AutowireScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, Data
                     dropdown.button(Text.translatable("redstone-helper.feature.auto_wire.modes.${mode.toString().lowercase()}")) {
                         dropdown.remove()
                         AutoWire.setMode(AutoWire.valueOf(mode.toString()))
-                        currentMode.text(Text.translatable("redstone-helper.feature.auto_wire.current_auto_wire_mode", Text.translatable("redstone-helper.feature.auto_wire.modes.${ClientOptions.autoWireMode.toString().lowercase()}")))
+                        currentMode.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_AUTOWIRE_CURRENTAUTOWIREMODE, Text.translatable("redstone-helper.feature.auto_wire.modes.${ClientOptions.autoWireMode.toString().lowercase()}")))
                         logger.debug("Selected new auto wire mode: " + ClientOptions.autoWireMode)
                     }
                 }