+++ src/main/kotlin/ua/pp/lumivoid/gui/BitsScreen.kt
@@ -4,6 +4,7 @@ import io.wispforest.owo.ui.base.BaseUIModelScreen
 import io.wispforest.owo.ui.component.CheckboxComponent
 import io.wispforest.owo.ui.component.LabelComponent
 import io.wispforest.owo.ui.container.FlowLayout
+import io.wispforest.owo.ui.core.Color
 import io.wispforest.owo.ui.core.Surface
 import net.minecraft.client.gui.widget.TextFieldWidget
 import net.minecraft.text.Text
@@ -15,6 +16,7 @@ import ua.pp.lumivoid.util.Calculate
 class BitsScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, DataSource.asset(Identifier.of(Constants.MOD_ID, "bits_ui_model"))) {
     private val logger = Constants.LOGGER
 
+    @Suppress("t")
     override fun build(rootComponent: FlowLayout) {
         logger.debug("Building AutowireScreen UI")
 
@@ -38,6 +40,18 @@ class BitsScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, DataSour
             layout.surface(Surface.DARK_PANEL)
         } else {
             layout.surface(Surface.PANEL)
+
+            val color: Color
+            Calculate.hexToRGB(0x3F3F3F).let { (r, g, b, a) ->
+                color = Color(r, g, b, a)
+            }
+            layout.children().forEach { component ->
+                logger.info("child")
+                if (component is LabelComponent && !component.text().string.contains("__colored")) {
+                    logger.info(component.text().string)
+                    component.color(color)
+                }
+            }
         }
         if (Config().enableBackgroundBlur) {
             rootComponent.surface(Surface.blur(100F, 10F))
@@ -45,9 +59,9 @@ class BitsScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, DataSour
             rootComponent.surface(Surface.VANILLA_TRANSLUCENT)
         }
 
-        toBitResult.text(Text.translatable("redstone-helper.feature.bits_operations.invalid_expression"))
-        fromBitResult.text(Text.translatable("redstone-helper.feature.bits_operations.invalid_expression"))
-        bitResult.text(Text.translatable("redstone-helper.feature.bits_operations.invalid_expression"))
+        toBitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_INVALIDEXPRESSION))
+        fromBitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_INVALIDEXPRESSION))
+        bitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_INVALIDEXPRESSION))
         toBitExpressionField.setMaxLength(999999999)
         bitExpressionField.setMaxLength(999999999)
         fromBitExpressionField.setMaxLength(999999999)
@@ -59,33 +73,33 @@ class BitsScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, DataSour
 
             try {
                 if (radix == 2) {
-                    toBitResult.text(Text.translatable("redstone-helper.feature.bits_operations.result", Integer.toBinaryString(toBitExpressionField.text.toInt())))
+                    toBitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_RESULT, Integer.toBinaryString(toBitExpressionField.text.toInt())))
                 } else {
-                    toBitResult.text(Text.translatable("redstone-helper.feature.bits_operations.result", "#" + Integer.toHexString(toBitExpressionField.text.toInt())))
+                    toBitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_RESULT, "#" + Integer.toHexString(toBitExpressionField.text.toInt())))
                 }
             } catch (e: NumberFormatException) {
                 logger.error("Not a number entered $e")
-                toBitResult.text(Text.translatable("redstone-helper.feature.bits_operations.invalid_expression"))
+                toBitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_INVALIDEXPRESSION))
             }
         }
         fromBitExpressionField.setChangedListener {
             logger.debug("Transform ${fromBitExpressionField.text} from bit")
 
             try {
-                fromBitResult.text(Text.translatable("redstone-helper.feature.bits_operations.result", Integer.parseInt(fromBitExpressionField.text, radix)))
+                fromBitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_RESULT, Integer.parseInt(fromBitExpressionField.text, radix)))
             } catch (e: NumberFormatException) {
-                fromBitResult.text(Text.translatable("redstone-helper.feature.bits_operations.invalid_expression"))
+                fromBitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_INVALIDEXPRESSION))
             }
         }
 
         hex.onChanged { checked ->
             if (checked) {
                 radix = 16
-                toBitText.text(Text.translatable("redstone-helper.feature.bits_operations.to_hex"))
-                fromBitText.text(Text.translatable("redstone-helper.feature.bits_operations.from_hex"))
+                toBitText.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_TOHEX))
+                fromBitText.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_FROMHEX))
             } else {
-                toBitText.text(Text.translatable("redstone-helper.feature.bits_operations.to_bit"))
-                fromBitText.text(Text.translatable("redstone-helper.feature.bits_operations.from_bit"))
+                toBitText.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_TOBIT))
+                fromBitText.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_FROMBIT))
                 radix = 2
             }
         }
@@ -97,9 +111,9 @@ class BitsScreen: BaseUIModelScreen<FlowLayout>(FlowLayout::class.java, DataSour
 
             val calcResult = Calculate.calc(Calculate.convertToRadix10(bitExpressionField.text))
             if (calcResult.isNaN()) {
-                bitResult.text(Text.translatable("redstone-helper.feature.bits_operations.invalid_expression"))
+                bitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_INVALIDEXPRESSION))
             } else {
-                bitResult.text(Text.translatable("redstone-helper.feature.bits_operations.bit_result", calcResult, Integer.toBinaryString(calcResult.toInt()), "#" + Integer.toHexString(calcResult.toInt())))
+                bitResult.text(Text.translatable(Constants.LOCALIZEIDS.FEATURE_BITSOPERATIONS_BITRESULT, calcResult, Integer.toBinaryString(calcResult.toInt()), "#" + Integer.toHexString(calcResult.toInt())))
             }
         }
     }