+++ src/main/kotlin/ua/pp/lumivoid/registration/PacketReceiverRegistration.kt
@@ -1,9 +1,6 @@
 package ua.pp.lumivoid.registration
 
-import ua.pp.lumivoid.network.receiver.server.ClearInventoryPacketReceiver
-import ua.pp.lumivoid.network.receiver.server.FillInventoryPacketReceiver
-import ua.pp.lumivoid.network.receiver.server.QuickTeleportPacketReceiver
-import ua.pp.lumivoid.network.receiver.server.SetBlockPacketReceiver
+import ua.pp.lumivoid.network.receiver.server.*
 
 object PacketReceiverRegistration {
     fun register() {
@@ -11,5 +8,6 @@ object PacketReceiverRegistration {
         QuickTeleportPacketReceiver.register()
         ClearInventoryPacketReceiver.register()
         FillInventoryPacketReceiver.register()
+        SetPlayerSlotInventoryReceiver.register()
     }
 }